Added a singletons.py file to regroup the global variables in one place. That's not pretty, but it could be better that what we had before. Also, I moved out the TestWindow class out of the __init__.py file
